+++ proc/mvm4/du.go
@@ -14,7 +14,7 @@ func newDecodeUnitWithBranchPredictor(bu *btbBranchUnit) *decodeUnit {
 	return &decodeUnit{bu: bu}
 }
 
-func (du *decodeUnit) cycle(currentCycle float32, app risc.Application, inBus comp.Bus[int], outBus comp.Bus[risc.InstructionRunner]) {
+func (du *decodeUnit) cycle(currentCycle int, app risc.Application, inBus comp.Bus[int], outBus comp.Bus[risc.InstructionRunner]) {
 	if du.pendingBranchResolution > 0 {
 		du.pendingBranchResolution--
 		return
